+++ angreal.toml
@@ -1,6 +1,5 @@
 author_name='Your name'
 author_email='Your email address (eq. you@example.com)'
-github_username='yourname'
 project_name='Name of the project (will be shown e.g. as the title in the readme)'
 project_slug='{{ project_name | lower | replace(from=" ", to="-") }}'
 package_name='{{ project_slug | replace(from="-", to="_") }}'
