+++ {{ project_slug }}/.angreal/task_setup.py
@@ -4,11 +4,11 @@ import os
 import subprocess
 
 
+dev = angreal.command_group(name="dev", about="commands for development tasks")
 
-
-@angreal.command(name='dev-setup', about="setup a development environment")
+@angreal.command(name='setup', about="setup a development environment")
 def setup_env():
-    one_up = os.path.join(angreal.get_root(), '..')    
+    one_up = os.path.join(angreal.get_root(), '..')
     return subprocess.run(
         (
             "mkdir -p .venv/angreal_python_dev;"
