--- {{ project_slug }}/.github/pull_request_template.md
@@ -1,10 +1,0 @@
-## Description
-
-A short description about the changes in this pull request. If the pull request is related to some issue, mention it
- here.
-
-## Checklist
-
-- [ ] Tests covering the new functionality have been added
-- [ ] Documentation has been updated OR the change is too minor to be documented
-- [ ] Changes are listed in the `CHANGELOG.md` OR changes are insignificant
