--- {{ project_slug }}/.github/workflows/docs.yaml
@@ -1,68 +1,0 @@
-# Sample workflow for building and deploying a Hugo site to GitHub Pages
-name: "Docs Deploy"
-
-on:
-  # Runs on pushes targeting the default branch
-  push:
-    branches: ["main"]
-
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
-permissions:
-  contents: read
-  pages: write
-  id-token: write
-
-# Allow one concurrent deployment
-concurrency:
-  group: "pages"
-  cancel-in-progress: true
-
-# Default to bash
-defaults:
-  run:
-    shell: bash
-
-jobs:
-  # Build job
-  build:
-    runs-on: ubuntu-latest
-    env:
-      HUGO_VERSION: 0.108.0
-    steps:
-      - name: Install Hugo CLI
-        run: |
-          {% raw %}wget -O ${{ runner.temp }}/hugo.deb https://github.com/gohugoio/hugo/releases/download/v${HUGO_VERSION}/hugo_${HUGO_VERSION}_linux-amd64.deb \
-          && sudo dpkg -i ${{ runner.temp }}/hugo.deb{% endraw %}
-      - name: Install Dart Sass Embedded
-        run: sudo snap install dart-sass-embedded
-      - name: Checkout
-        uses: actions/checkout@v3
-        with: 
-          submodules: true  # Fetch Hugo themes (true OR recursive)
-          fetch-depth: 0   
-      - name: Setup Pages
-        id: pages
-        uses: actions/configure-pages@v2
-      - name: Install Node.js dependencies
-        run: "[[ -f package-lock.json || -f npm-shrinkwrap.json ]] && npm ci || true"
-      - name: Build with Hugo
-        run: cd docs/ && hugo --minify
-      - name: Upload artifact
-        uses: actions/upload-pages-artifact@v1
-        with:
-          path: ./docs/public
-
-  # Deployment job
-  deploy:
-    environment:
-      name: github-pages
-      url: {% raw %}${{ steps.deployment.outputs.page_url }}{% endraw %}
-    runs-on: ubuntu-latest
-    needs: build
-    steps:
-      - name: Deploy to GitHub Pages
-        id: deployment
-        uses: actions/deploy-pages@v1
