--- {{ project_slug }}/.github/workflows/publish.yaml
@@ -1,40 +1,0 @@
-name: Release
-on: push
-concurrency:
-  group: {% raw %}${{ github.workflow }}-${{ github.ref_name }}-${{ github.event.pull_request.number || github.sha }}{% endraw %}
-  cancel-in-progress: true
-jobs:
-  build-n-publish:
-    name: Build and publish Python 🐍 distributions 📦 to PyPI and TestPyPI
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@master
-    - name: Set up Python 3.10
-      uses: actions/setup-python@v3
-      with:
-        python-version: "3.10"
-    - name: Install pypa/build
-      run: >-
-        python -m
-        pip install
-        build
-        --user
-    - name: Build a binary wheel and a source tarball
-      run: >-
-        python -m
-        build
-        --sdist
-        --wheel
-        --outdir dist/
-        .
-    - name: Publish distribution 📦 to Test PyPI
-      if: startsWith(github.ref, 'refs/tags')
-      uses: pypa/gh-action-pypi-publish@release/v1
-      with:
-        password: {% raw %}${{ secrets.TEST_PYPI_API_TOKEN }}{% endraw %}
-        repository_url: https://test.pypi.org/legacy/
-    - name: Publish distribution 📦 to PyPI
-      if: startsWith(github.ref, 'refs/tags')
-      uses: pypa/gh-action-pypi-publish@release/v1
-      with:
-        password: {% raw %}${{ secrets.PYPI_API_TOKEN }}{% endraw %}
