--- {{ project_slug }}/.github/workflows/test.yaml
@@ -1,85 +1,0 @@
-name: Testing
-
-on:
-  push: 
-    branches: [main]
-  pull_request:
-    branches: [main]
-concurrency:
-  group: {% raw %}${{ github.workflow }}-${{ github.ref_name }}-${{ github.event.pull_request.number || github.sha }}{% endraw %}
-  cancel-in-progress: true
-env:
-  FORCE_COLOR: "1" # Make tools pretty.
-  PIP_DISABLE_PIP_VERSION_CHECK: "1"
-  PIP_NO_PYTHON_VERSION_WARNING: "1"
-  SETUPTOOLS_SCM_PRETEND_VERSION: "1.0" # avoid warnings about shallow checkout
-  PYTHON_LATEST: "3.11"
-
-
-jobs:
-  tests:
-    name: tests on {% raw %}${{ matrix.python-version }}{% endraw %}
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: true
-      matrix:
-        python-version: 
-          - "3.7"
-          - "3.8"
-          - "3.9"
-          - "3.10"
-          - "3.11"
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        with:
-          python-version: {% raw %}${{ matrix.python-version }}{% endraw %}
-        
-      - name: unit tests
-        run: |
-          python -VV
-          python -m site
-          python -m pip install --upgrade pip setuptools wheel angreal
-          angreal dev-setup
-          angreal run-tests
-  mypy:
-    name: mypy on {% raw %}${{ matrix.python-version }}{% endraw %}
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix: 
-        python-version:
-          - "3.7"
-          - "3.8"
-          - "3.9"
-          - "3.10"
-          - "3.11"
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        with:
-          python-version: {% raw %}${{ matrix.python-version }}{% endraw %}
-      - name: static tests 
-        run: |
-          pip install .[dev]
-          angreal dev-setup
-          angreal static-analysis
-
-  install-dev:
-    name: Verify dev env
-    runs-on: {% raw %}${{ matrix.os }}{% endraw %}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        with:
-          python-version: {% raw %}${{env.PYTHON_LATEST}}{% endraw %}
-      - name: install and test import
-        run: |
-          python -m pip install -e .
-          python -c 'import {{ package_name }}; print({{ package_name }}.__version__)'
-
-
-
