+++ {{ project_slug }}/README.md
@@ -1,19 +1,5 @@
 # {{ project_name }}
 
-[![PyPI](https://img.shields.io/pypi/v/{{ project_slug }}?style=flat-square)](https://pypi.python.org/pypi/{{ project_slug }}/)
-[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/{{ project_slug}}?style=flat-square)](https://pypi.python.org/pypi/{{ project_slug }}/)
-[![PyPI - License](https://img.shields.io/pypi/l/{{ project_slug }}?style=flat-square)](https://pypi.python.org/pypi/{{ project_slug }}/)
-
-
----
-
-**Documentation**: [https://{{ github_username }}.github.io/{{ project_slug}}](https://{{ github_username }}.github.io/{{ project_slug}})
-
-**Source Code**: [https://github.com/{{ github_username }}/{{ project_slug }}](https://github.com/{{ github_username }}/{{ project_slug }})
-
-**PyPI**: [https://pypi.org/project/{{ project_slug }}/](https://pypi.org/project/{{ project_slug }}/)
-
----
 
 {{ project_name }} is going to be the next big thing !
 
@@ -31,22 +17,6 @@ pip install {{ project_slug }}
 * Requirements:
   * `angreal`
 * `pip install angreal && angreal setup`
-
-### Testing
-
-```sh
-angreal run-tests --help
-```
-
-### Documentation
-
-
-
-### Releasing
-
-
-
-### Pre-commit
 
 
 This project was generated using the [angreal python template](https://github.com/angreal/python) template.
