+++ {{ project_slug }}/setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = {{ project_slug }}
 version = attr: {{ package_name }}.__version__
-url = "https://github.com/{{ github_username }}/{{ project_slug }}"
+url = ""
 author = {{ author_name }}
 author_email = {{ author_email }}
 description = {{ project_short_description }}
